requirements.txt
pefile>=2024.8.26
yara-python>=4.5.0
oletools>=0.60.0
httpx>=0.28.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
rich>=14.0.0
